Starting relay Mission

+++ Projects/CoX/Common/GameData/Contact.h
@@ -82,12 +82,12 @@ public:
        QString         m_name;
        QString         m_location_description;
        uint32_t        m_npc_id;
-       uint8_t         m_contact_idx;
-       uint8_t         m_current_standing;
-       uint8_t         m_confidant_threshold;
-       uint8_t         m_friend_threshold;
-       uint8_t         m_complete_threshold;
-       uint8_t         m_task_index            = 0;
+       uint32_t        m_contact_idx;
+       uint32_t        m_current_standing;
+       uint32_t        m_confidant_threshold;
+       uint32_t        m_friend_threshold;
+       uint32_t        m_complete_threshold;
+       uint32_t        m_task_index            = 0;
        bool            m_notify_player         = false;
        bool            m_can_use_cell          = false;
        bool            m_has_location          = false;
